mvp/unlisted_jobs/requirements.txt
# Core dependencies
psycopg2-binary>=2.9.0
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Data acquisition
requests>=2.26.0
beautifulsoup4>=4.10.0
lxml>=4.6.0
openpyxl>=3.0.0  # For reading Excel files (H-1B data)

# Geographic data
pgeocode>=0.4.0
geopy>=2.2.0

# Statistics and modeling (for salary/headcount models)
scikit-learn>=1.0.0
statsmodels>=0.13.0

# Optional: Bayesian modeling
# pymc3>=3.11.0  # For hierarchical Bayesian salary model

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0

# Utilities
python-dotenv>=0.19.0
tqdm>=4.62.0

# Logging
colorlog>=6.6.0
